+++ changelog.md
@@ -1,5 +1,10 @@
 # Change Log for GuestConfiguration module
 
+## v1.19.0
+
+- Resolve bug in custom categories
+- Catch signing error output when certificate is invalid
+
 ## v1.18.0
 
 - seperated module and tests from base repo